libmctp: usb: Fix debug buf print infinite loop
Debug print was going in an infinite loop since iterator was incorrectly defined
as a uint8_t, but with batch Tx our packet sizes are larger

- debug message can be removed once batch_tx issues are resolved

Fixes jira https://jirasw.nvidia.com/browse/DGXOPENBMC-9471

+++ usb.c
@@ -339,7 +339,7 @@ void mctp_send_tx_queue_usb(struct mctp_bus *bus)
 
 			mctp_prinfo("buffer: ");
 			dt = (unsigned char *)usb->txbuf;
-			for (uint8_t i=0; i< usb_buf_len; i++){
+			for (unsigned int i=0; i< usb_buf_len; i++){
 				mctp_prinfo("%02X ", (unsigned int) dt[i]);
 			}
 
@@ -364,7 +364,7 @@ void mctp_send_tx_queue_usb(struct mctp_bus *bus)
 
 	mctp_prinfo("buffer outside: ");
 	dt = (unsigned char *)usb->txbuf;
-	for (uint8_t i=0; i< usb_buf_len; i++){
+	for (unsigned int i=0; i< usb_buf_len; i++){
 				mctp_prinfo("%02X ", (unsigned int) dt[i]);
 	}
 	rv = mctp_usb_tx(usb, usb_buf_len);
